combine authorization service with user
