+++ .gitignore
@@ -1,6 +1,6 @@
 HELP.md
 target/
-!.mvn/wrapper/**
+!.mvn/wrapper/
 !mvnw
 !mvnw.cmd
 !HELP.md
