+++ OAuth2/src/main/java/com/wolfhack/cloud/controller/UserRestController.java
@@ -3,11 +3,8 @@ package com.wolfhack.cloud.controller;
 import com.wolfhack.cloud.dto.UserCreationDTO;
 import com.wolfhack.cloud.dto.UserResponseDTO;
 import com.wolfhack.cloud.dto.mapper.UserMapper;
-import com.wolfhack.cloud.exception.UserExistsException;
-import com.wolfhack.cloud.exception.UserNotFoundException;
 import com.wolfhack.cloud.handler.error.ErrorBody;
 import com.wolfhack.cloud.handler.error.ValidationErrorBody;
-import com.wolfhack.cloud.model.User;
 import com.wolfhack.cloud.service.UserService;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.media.Content;
