+++ OAuth2/src/main/resources/application.properties
@@ -3,5 +3,13 @@ server.port=8081
 #eureka
 eureka.client.service-url.defaultZone=http://eureka-server:8761/eureka
 eureka.instance.prefer-ip-address=true
+#mysql
+spring.jpa.hibernate.ddl-auto=update
+spring.datasource.url=jdbc:mysql://mysql-eshop:3306/user
+spring.datasource.username=root
+spring.datasource.password=WolFHacK702
+spring.main.allow-bean-definition-overriding=true
+#flyway
+management.endpoint.flyway.enabled=false
 #actuators
 management.endpoints.web.exposure.include=*
