+++ ProductService/pom.xml
@@ -59,6 +59,21 @@
             <artifactId>spring-boot-starter-cache</artifactId>
             <version>2.4.0</version>
         </dependency>
+
+        <!--    DataBase    -->
+        <dependency>
+            <groupId>org.flywaydb</groupId>
+            <artifactId>flyway-mysql</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter-data-jpa</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>mysql</groupId>
+            <artifactId>mysql-connector-java</artifactId>
+            <scope>runtime</scope>
+        </dependency>
         <!--    Test    -->
         <dependency>
             <groupId>org.springframework.boot</groupId>
