--- UserService/Dockerfile-dev
@@ -1,4 +1,0 @@
-FROM adoptopenjdk/openjdk11:latest
-ARG JAR_FILE=target/*.jar
-COPY ${JAR_FILE} user-service.jar
-ENTRYPOINT ["java", "-jar", "user-service.jar"]
