--- UserService/src/main/java/com/wolfhack/cloud/model/Role.java
@@ -1,11 +1,0 @@
-package com.wolfhack.cloud.model;
-
-public enum Role {
-
-    ADMIN, USER;
-
-    @Override
-    public String toString() {
-        return "ROLE_"+name();
-    }
-}
