--- UserService/src/main/java/com/wolfhack/cloud/model/User.java
@@ -1,74 +1,0 @@
-package com.wolfhack.cloud.model;
-
-import lombok.Getter;
-import lombok.Setter;
-
-import javax.persistence.*;
-import java.sql.Date;
-import java.util.Objects;
-
-@Entity(name = "users")
-@Getter @Setter
-public class User {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "user_id")
-    private Long id;
-
-    @Column(name = "login")
-    private String login;
-
-    @Column(name = "username")
-    private String username;
-
-    @Column(name = "password")
-    private String password;
-
-    @Column(name = "photo")
-    private String photo;
-
-    @Column(name = "active")
-    private boolean active;
-
-    @Column(name = "activation_code")
-    private String activationCode;
-
-    @Column(name = "first_name")
-    private String firstName;
-
-    @Column(name = "second_name")
-    private String secondName;
-
-    @Column(name = "middle_name")
-    private String middleName;
-
-    @Column(name = "phone")
-    private String phone;
-
-    @Column(name = "city")
-    private String city;
-
-    @Column(name = "register_date")
-    private Date registerDate;
-
-    @Enumerated(EnumType.STRING)
-    private Role role;
-
-    public boolean isAdmin() {
-        return Objects.equals(role.toString(), "ADMIN");
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        User user = (User) o;
-        return login.equals(user.login);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(id);
-    }
-}
