--- UserService/src/main/java/com/wolfhack/cloud/repository/UserRepository.java
@@ -1,14 +1,0 @@
-package com.wolfhack.cloud.repository;
-
-import com.wolfhack.cloud.model.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-
-    Optional<User> findByLogin(String login);
-    boolean existsUserByLogin(String login);
-}
