--- UserService/src/main/resources/application.properties
@@ -1,14 +1,0 @@
-#spring server
-spring.application.name=user-service
-server.port=8085
-#eureka
-eureka.client.service-url.defaultZone=http://eureka-server:8761/eureka
-eureka.instance.prefer-ip-address=true
-#mysql
-spring.jpa.hibernate.ddl-auto=update
-spring.datasource.url=jdbc:mysql://mysql-eshop:3306/user
-spring.datasource.username=root
-spring.datasource.password=WolFHacK702
-spring.main.allow-bean-definition-overriding=true
-
-management.endpoint.flyway.enabled=false
