--- UserService/src/main/resources/banner.txt
@@ -1,7 +1,0 @@
-
-
-,--. ,--.                          ,---.                              ,--.
-|  | |  |  ,---.   ,---.  ,--.--. '   .-'   ,---.  ,--.--. ,--.  ,--. `--'  ,---.  ,---.
-|  | |  | (  .-'  | .-. : |  .--' `.  `-.  | .-. : |  .--'  \  `'  /  ,--. | .--' | .-. :
-'  '-'  ' .-'  `) \   --. |  |    .-'    | \   --. |  |      \    /   |  | \ `--. \   --.
- `-----'  `----'   `----' `--'    `-----'   `----' `--'       `--'    `--'  `---'  `----'
